backend/requirements.txt
Django>=4.0,<6.0
djangorestframework
django-environ
PyMySQL
mysqlclient
djangorestframework-simplejwt
django-cors-headers
weasyprint
django-extensions
cryptography
